Modify Noise type
In reader.py, change the way to get the vals. First read the legend and then get the whole table. From the legend get the right columns of the table to be displayed.
Fix the refresh method by removing everything including the legend and reading everythong again. When a file or the selection change, refresh everything each time. The legend is then refreshed too.
When switching to a noise type that has not been found in a data table, then add stars to the legend of the curve to indicate that it is not present on the graph