update pyqtgraph to 0.12.3
pyqtgraph 0.10.0 doesn't work anymore in a modern Python installation,
since some PyQt functions such as setAlpha() expect an int but are
being called with a float.

This updates pyqtgraph to 0.12.3, works correctly. Versions at or
above 0.13.0 don't work anymore with PyBER, since there are some
changes in the pyqtgraph API.